+++ scrapyrt/exceptions.py
@@ -1,4 +1,0 @@
-class ScrapyrtDeprecationWarning(Warning):
-    """Warning category for deprecated features, since the default
-    DeprecationWarning is silenced on Python 2.7+.
-    """